refactor(ui): one typing indicator — canonical TurnStatus in chat-typing-indicator (#12188 Phase 3)
Move the overlay's phase-aware TurnStatusInner (dots + debounced phase label)
into composites/chat/chat-typing-indicator.tsx as the canonical TurnStatus;
all three dot triads (default / game-modal / turn-status) now render through
one shared TypingDots primitive. Drop the dead TypingIndicator.agentAvatarSrc
prop (accepted but never read) and its ChatView call site.
