+++ packages/ui/src/components/shell/ContinuousChatOverlay.tsx
@@ -96,6 +96,7 @@ import { findFollowupsRegions } from "../chat/message-followups-parser";
 import { findFormRegions } from "../chat/message-form-parser";
 import { ThinkingBlock } from "../chat/ThinkingBlock";
 import { withTranscriptMarker } from "../chat/TranscriptViewerOverlay";
+import { TurnStatus } from "../composites/chat/chat-typing-indicator";
 import { Button } from "../ui/button";
 import { Input } from "../ui/input";
 import { Textarea } from "../ui/textarea";
@@ -620,151 +621,10 @@ function PillHandle({
   );
 }
 
-/** Humanize an action/tool symbol ("SEND_MESSAGE" → "Send message") for the
- *  status label so it reads as prose, not a constant. */
-function humanizeStatusName(name: string): string {
-  const cleaned = name.trim().replace(/[_-]+/g, " ").toLowerCase();
-  if (!cleaned) return "";
-  return cleaned.charAt(0).toUpperCase() + cleaned.slice(1);
-}
-
-/** The phase label shown beside the breathing dots. Server-provided `label`
- *  always wins; otherwise derive a concise phrase from the kind (+ action/tool
- *  name when present). */
-function turnStatusLabel(status: ChatTurnStatus): string {
-  if (status.label?.trim()) return status.label.trim();
-  switch (status.kind) {
-    case "thinking":
-      return "Thinking";
-    case "streaming":
-      return "Replying";
-    case "running_action":
-      return status.actionName
-        ? `Running ${humanizeStatusName(status.actionName)}`
-        : "Working";
-    case "running_tool":
-      return status.toolName
-        ? `Using ${humanizeStatusName(status.toolName)}`
-        : "Using a tool";
-    case "evaluating":
-      return "Reflecting";
-    case "waking":
-      return "Waking the agent";
-    case "speaking":
-      return "Speaking";
-  }
-}
-
-// Min time (ms) a phase label must stay on screen before the next one replaces
-// it. Without this, a fast thinking→action→streaming sequence flickers through
-// labels faster than the eye can read. The text content the user sees is
-// debounced; the dots animate continuously throughout.
-const STATUS_MIN_DWELL_MS = 320;
-
-/** Debounce a fast-changing status to a min on-screen dwell so a rapid
- *  thinking→action→streaming sequence doesn't strobe the label. The dots animate
- *  continuously; only the words wait their turn. Returns null when status is
- *  null (between turns). */
-function useDebouncedTurnStatus(
-  status: ChatTurnStatus | null,
-): ChatTurnStatus | null {
-  const [shown, setShown] = React.useState<ChatTurnStatus | null>(status);
-  const lastChangeRef = React.useRef(0);
-  const timerRef = React.useRef<number | null>(null);
-  React.useEffect(() => {
-    if (!status) {
-      setShown(null);
-      return;
-    }
-    const now = Date.now();
-    const elapsed = now - lastChangeRef.current;
-    if (elapsed >= STATUS_MIN_DWELL_MS) {
-      lastChangeRef.current = now;
-      setShown(status);
-      return;
-    }
-    if (timerRef.current !== null) window.clearTimeout(timerRef.current);
-    timerRef.current = window.setTimeout(() => {
-      lastChangeRef.current = Date.now();
-      setShown(status);
-      timerRef.current = null;
-    }, STATUS_MIN_DWELL_MS - elapsed);
-    return () => {
-      if (timerRef.current !== null) {
-        window.clearTimeout(timerRef.current);
-        timerRef.current = null;
-      }
-    };
-  }, [status]);
-  return shown;
-}
-
-/**
- * The breathing dots + phase label content of the status indicator, WITHOUT a
- * bubble/motion wrapper — used both standalone (wrapped by TurnStatusIndicator)
- * and inline inside the in-flight assistant bubble.
- *
- * Mirrors ChatVoiceStatusBar's a11y (`role="status"` + `aria-live="polite"`).
- * Honors reduced motion (no pulse). Brand-safe: orange (the accent) ONLY for
- * `speaking`; every other phase is neutral white. No blue anywhere. Degrades to
- * plain dots when no status has arrived.
- */
-function TurnStatusInner({
-  status,
-  showLabel = true,
-}: {
-  status: ChatTurnStatus | null;
-  showLabel?: boolean;
-}): React.JSX.Element {
-  const shown = useDebouncedTurnStatus(status);
-  const speaking = shown?.kind === "speaking";
-  const label =
-    showLabel && shown && shown.kind !== "thinking"
-      ? turnStatusLabel(shown)
-      : null;
-  return (
-    <span
-      className="inline-flex items-center gap-2"
-      data-testid="turn-status-indicator"
-      data-status-kind={shown?.kind ?? "none"}
-      role="status"
-      aria-live="polite"
-    >
-      <span
-        className="flex gap-1.5"
-        data-testid="typing-dots"
-        aria-hidden="true"
-      >
-        {[0, 1, 2].map((i) => (
-          <span
-            key={i}
-            className={cn(
-              "h-1.5 w-1.5 animate-pulse rounded-full motion-reduce:animate-none",
-              speaking ? "bg-[rgba(255,190,140,0.9)]" : "bg-white/70",
-            )}
-            style={{ animationDelay: `${i * 180}ms` }}
-          />
-        ))}
-      </span>
-      {label ? (
-        <span
-          className={cn(
-            "text-[13px] font-medium",
-            speaking ? "text-[rgba(255,200,150,0.95)]" : "text-white/90",
-          )}
-          data-testid="turn-status-label"
-        >
-          {label}
-        </span>
-      ) : null}
-    </span>
-  );
-}
-
 /**
  * The rich, phase-aware status row shown while the assistant works (#8813),
- * replacing the bare typing dots in the pre-placeholder gap. Wraps
- * TurnStatusInner in its own glass bubble + fade so it reads as a turn.
+ * replacing the bare typing dots in the pre-placeholder gap. Wraps the
+ * canonical TurnStatus in its own glass bubble + fade so it reads as a turn.
  */
 function TurnStatusIndicator({
   status,
@@ -794,7 +654,7 @@ function TurnStatusIndicator({
           speaking ? "border-[rgba(255,180,120,0.45)]" : "border-white/10",
         )}
       >
-        <TurnStatusInner status={status} />
+        <TurnStatus status={status} />
       </div>
     </motion.div>
   );
@@ -1409,7 +1269,7 @@ const ThreadLine = React.memo(function ThreadLine({
             // in the standalone status row so the bubble never flashes
             // "Running …" text in place of the answer.
             <>
-              <TurnStatusInner status={turnStatus ?? null} showLabel={false} />
+              <TurnStatus status={turnStatus ?? null} showLabel={false} />
               {message.attachments?.length ? (
                 <MessageAttachments attachments={message.attachments} />
               ) : null}
